Better testing of email in PersonTest
Adds tests for the email to check for random domain names containing
lower case letters and numbers differing in size. Also checks all ISO
3166 registered 2 letter cTLD (Country Top Level Domains) for vaild
emails, and all other 2 letter combination for invalid emails.

For an updated version of the cTLD list given in the task, which only
includes 233 of the now 249 ISO 3166 registered 2 letter cTLDs see this
link:
http://pastebin.com/chG6WLWF